feat:画像をbase64形式に変換する処理
- markdownEditorに埋め込んだファイルをbase64形式に変換する関数を作った
- MarkDownEditor.tsxの可読性向上のため、いくつかの関数をutilsファイルに移動させた
