+++ src/app/_utils/markdown_editor/ImageFnc/fileToBase64.ts
@@ -1,0 +1,9 @@
+export const fileToBase64 = (file: File) => {
+  if (file) {
+    const reader = new FileReader();
+    reader.onload = () => {
+      console.log(reader.result);
+    };
+    return reader.readAsDataURL(file);
+  }
+};
